Multi batch version 3
