+++ exporter/exporterhelper/internal/queuebatch/default_batcher.go
@@ -2,45 +2,25 @@
 // SPDX-License-Identifier: Apache-2.0
 
 package queuebatch // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal/queuebatch"
-
 import (
 	"context"
 	"sync"
-	"time"
-
-	"go.uber.org/multierr"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/request"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/sender"
 )
 
-type batch struct {
-	ctx  context.Context
-	req  request.Request
-	done multiDone
-}
-
-type batcherSettings[T any] struct {
-	sizerType   request.SizerType
-	sizer       request.Sizer[T]
-	partitioner Partitioner[T]
-	next        sender.SendFunc[T]
-	maxWorkers  int
-}
-
-// defaultBatcher continuously batch incoming requests and flushes asynchronously if minimum size limit is met or on timeout.
 type defaultBatcher struct {
-	cfg            BatchConfig
-	workerPool     *chan struct{}
-	sizerType      request.SizerType
-	sizer          request.Sizer[request.Request]
-	consumeFunc    sender.SendFunc[request.Request]
-	stopWG         sync.WaitGroup
-	currentBatchMu sync.Mutex
-	currentBatch   *batch
-	timer          *time.Timer
-	shutdownCh     chan struct{}
+	cfg         BatchConfig
+	workerPool  *chan struct{}
+	sizerType   request.SizerType
+	sizer       request.Sizer[request.Request]
+	partitioner Partitioner[request.Request]
+	consumeFunc sender.SendFunc[request.Request]
+
+	shardMapMu sync.Mutex
+	shards     map[string]*shard
 }
 
 func newDefaultBatcher(bCfg BatchConfig, bSet batcherSettings[request.Request]) *defaultBatcher {
@@ -57,209 +37,58 @@ func newDefaultBatcher(bCfg BatchConfig, bSet batcherSettings[request.Request]) 
 		workerPool:  &workerPool,
 		sizerType:   bSet.sizerType,
 		sizer:       bSet.sizer,
+		partitioner: bSet.partitioner,
 		consumeFunc: bSet.next,
-		stopWG:      sync.WaitGroup{},
-		shutdownCh:  make(chan struct{}, 1),
+		shardMapMu:  sync.Mutex{},
+		shards:      make(map[string]*shard),
 	}
 }
 
-func (qb *defaultBatcher) resetTimer() {
-	if qb.cfg.FlushTimeout > 0 {
-		qb.timer.Reset(qb.cfg.FlushTimeout)
+func (qb *defaultBatcher) getShard(ctx context.Context, req request.Request) *shard {
+	key := ""
+	if qb.partitioner != nil {
+		key = qb.partitioner.GetKey(ctx, req)
 	}
+	qb.shardMapMu.Lock()
+	defer qb.shardMapMu.Unlock()
+
+	s, ok := qb.shards[key]
+	if !ok {
+		s = &shard{
+			cfg:         qb.cfg,
+			workerPool:  qb.workerPool,
+			sizerType:   qb.sizerType,
+			sizer:       qb.sizer,
+			consumeFunc: qb.consumeFunc,
+			stopWG:      sync.WaitGroup{},
+			shutdownCh:  make(chan struct{}, 1),
+		}
+		qb.shards[key] = s
+		_ = s.Start(ctx, nil)
+	}
+	return s
+}
+
+func (qb *defaultBatcher) Start(_ context.Context, _ component.Host) error {
+	return nil
 }
 
 func (qb *defaultBatcher) Consume(ctx context.Context, req request.Request, done Done) {
-	qb.currentBatchMu.Lock()
-
-	if qb.currentBatch == nil {
-		reqList, mergeSplitErr := req.MergeSplit(ctx, int(qb.cfg.MaxSize), qb.sizerType, nil)
-		if mergeSplitErr != nil || len(reqList) == 0 {
-			done.OnDone(mergeSplitErr)
-			qb.currentBatchMu.Unlock()
-			return
-		}
-
-		// If more than one flush is required for this request, call done only when all flushes are done.
-		if len(reqList) > 1 {
-			done = newRefCountDone(done, int64(len(reqList)))
-		}
-
-		// We have at least one result in the reqList. Last in the list may not have enough data to be flushed.
-		// Find if it has at least MinSize, and if it does then move that as the current batch.
-		lastReq := reqList[len(reqList)-1]
-		if qb.sizer.Sizeof(lastReq) < qb.cfg.MinSize {
-			// Do not flush the last item and add it to the current batch.
-			reqList = reqList[:len(reqList)-1]
-			qb.currentBatch = &batch{
-				ctx:  ctx,
-				req:  lastReq,
-				done: multiDone{done},
-			}
-			qb.resetTimer()
-		}
-
-		qb.currentBatchMu.Unlock()
-		for i := 0; i < len(reqList); i++ {
-			qb.flush(ctx, reqList[i], done)
-		}
-
-		return
-	}
-
-	reqList, mergeSplitErr := qb.currentBatch.req.MergeSplit(ctx, int(qb.cfg.MaxSize), qb.sizerType, req)
-	// If failed to merge signal all Done callbacks from current batch as well as the current request and reset the current batch.
-	if mergeSplitErr != nil || len(reqList) == 0 {
-		done.OnDone(mergeSplitErr)
-		qb.currentBatchMu.Unlock()
-		return
-	}
-
-	// If more than one flush is required for this request, call done only when all flushes are done.
-	if len(reqList) > 1 {
-		done = newRefCountDone(done, int64(len(reqList)))
-	}
-
-	// We have at least one result in the reqList, if more results here is what that means:
-	// - First result will contain items from the current batch + some results from the current request.
-	// - All other results except first will contain items only from current request.
-	// - Last result may not have enough data to be flushed.
-
-	// Logic on how to deal with the current batch:
-	qb.currentBatch.req = reqList[0]
-	qb.currentBatch.done = append(qb.currentBatch.done, done)
-	qb.currentBatch.ctx = contextWithMergedLinks(qb.currentBatch.ctx, ctx)
-
-	// Save the "currentBatch" if we need to flush it, because we want to execute flush without holding the lock, and
-	// cannot unlock and re-lock because we are not done processing all the responses.
-	var firstBatch *batch
-	// Need to check the currentBatch if more than 1 result returned or if 1 result return but larger than MinSize.
-	if len(reqList) > 1 || qb.sizer.Sizeof(qb.currentBatch.req) >= qb.cfg.MinSize {
-		firstBatch = qb.currentBatch
-		qb.currentBatch = nil
-	}
-	// At this moment we dealt with the first result which is iter in the currentBatch or in the `firstBatch` we will flush.
-	reqList = reqList[1:]
-
-	// If we still have results to process, then we need to check if the last result has enough data to flush, or we add it to the currentBatch.
-	if len(reqList) > 0 {
-		lastReq := reqList[len(reqList)-1]
-		if qb.sizer.Sizeof(lastReq) < qb.cfg.MinSize {
-			// Do not flush the last item and add it to the current batch.
-			reqList = reqList[:len(reqList)-1]
-			qb.currentBatch = &batch{
-				ctx:  ctx,
-				req:  lastReq,
-				done: multiDone{done},
-			}
-			qb.resetTimer()
-		}
-	}
-
-	qb.currentBatchMu.Unlock()
-	if firstBatch != nil {
-		qb.flush(firstBatch.ctx, firstBatch.req, firstBatch.done)
-	}
-	for i := 0; i < len(reqList); i++ {
-		qb.flush(ctx, reqList[i], done)
-	}
+	shard := qb.getShard(ctx, req)
+	shard.Consume(ctx, req, done)
 }
 
-// startTimeBasedFlushingGoroutine starts a goroutine that flushes on timeout.
-func (qb *defaultBatcher) startTimeBasedFlushingGoroutine() {
-	qb.stopWG.Add(1)
-	go func() {
-		defer qb.stopWG.Done()
-		for {
-			select {
-			case <-qb.shutdownCh:
-				return
-			case <-qb.timer.C:
-				qb.flushCurrentBatchIfNecessary()
-			}
-		}
-	}()
-}
-
-// Start starts the goroutine that reads from the queue and flushes asynchronously.
-func (qb *defaultBatcher) Start(_ context.Context, _ component.Host) error {
-	if qb.cfg.FlushTimeout > 0 {
-		qb.timer = time.NewTimer(qb.cfg.FlushTimeout)
-		qb.startTimeBasedFlushingGoroutine()
+func (qb *defaultBatcher) Shutdown(ctx context.Context) error {
+	qb.shardMapMu.Lock()
+	defer qb.shardMapMu.Unlock()
+	stopWG := sync.WaitGroup{}
+	for _, shard := range qb.shards {
+		stopWG.Add(1)
+		go func() {
+			_ = shard.Shutdown(ctx)
+			stopWG.Done()
+		}()
 	}
-
+	stopWG.Wait()
 	return nil
-}
-
-// flushCurrentBatchIfNecessary sends out the current request batch if it is not nil
-func (qb *defaultBatcher) flushCurrentBatchIfNecessary() {
-	qb.currentBatchMu.Lock()
-	if qb.currentBatch == nil {
-		qb.currentBatchMu.Unlock()
-		return
-	}
-	batchToFlush := qb.currentBatch
-	qb.currentBatch = nil
-	qb.currentBatchMu.Unlock()
-
-	// flush() blocks until successfully started a goroutine for flushing.
-	qb.flush(batchToFlush.ctx, batchToFlush.req, batchToFlush.done)
-	qb.resetTimer()
-}
-
-// flush starts a goroutine that calls consumeFunc. It blocks until a worker is available if necessary.
-func (qb *defaultBatcher) flush(ctx context.Context, req request.Request, done Done) {
-	qb.stopWG.Add(1)
-	if qb.workerPool != nil {
-		<-*qb.workerPool
-	}
-	go func() {
-		defer qb.stopWG.Done()
-		done.OnDone(qb.consumeFunc(ctx, req))
-		if qb.workerPool != nil {
-			*qb.workerPool <- struct{}{}
-		}
-	}()
-}
-
-// Shutdown ensures that queue and all Batcher are stopped.
-func (qb *defaultBatcher) Shutdown(_ context.Context) error {
-	close(qb.shutdownCh)
-	// Make sure execute one last flush if necessary.
-	qb.flushCurrentBatchIfNecessary()
-	qb.stopWG.Wait()
-	return nil
-}
-
-type multiDone []Done
-
-func (mdc multiDone) OnDone(err error) {
-	for _, d := range mdc {
-		d.OnDone(err)
-	}
-}
-
-type refCountDone struct {
-	done     Done
-	mu       sync.Mutex
-	refCount int64
-	err      error
-}
-
-func newRefCountDone(done Done, refCount int64) Done {
-	return &refCountDone{
-		done:     done,
-		refCount: refCount,
-	}
-}
-
-func (rcd *refCountDone) OnDone(err error) {
-	rcd.mu.Lock()
-	defer rcd.mu.Unlock()
-	rcd.err = multierr.Append(rcd.err, err)
-	rcd.refCount--
-	if rcd.refCount == 0 {
-		// No more references, call done.
-		rcd.done.OnDone(rcd.err)
-	}
 }
