+++ exporter/exporterhelper/internal/queuebatch/queue_batch.go
@@ -76,13 +76,6 @@ func newQueueBatch(
 			// If partitioning is not enabled or if paritition is done at queue level, we can use the default batcher.
 			if set.Partitioner == nil {
 				b = newDefaultBatcher(*cfg.Batch, batcherSettings[request.Request]{
-					sizerType:  cfg.Sizer,
-					sizer:      sizer,
-					next:       next,
-					maxWorkers: cfg.NumConsumers,
-				})
-			} else {
-				b = newMultiBatcher(*cfg.Batch, batcherSettings[request.Request]{
 					sizerType:   cfg.Sizer,
 					sizer:       sizer,
 					partitioner: set.Partitioner,
